requirements.txt
# Base Dependencies
numpy>=1.9.0
scipy>=0.16.0
scikit-learn>=0.15
patsy>=0.4
sphinx==1.3.1
pandas==0.17.0
future==0.15.2

# Extensions
git+https://github.com/jmetzen/sparse-filtering.git
git+https://github.com/all-umass/metric-learn.git
git+https://github.com/nicodv/kmodes.git
git+https://github.com/amueller/patsylearn.git
git+https://github.com/predikto/pyculearn.git

lda>=1.0.2
sklearn-compiledtrees>=1.2